requirements.txt
covid-19
covid
tweepy
pandas
